+++ test/UnitTests.js
@@ -288,6 +288,13 @@ describe('External', function() {
 
     it('should call _request with the correct parameters', function() {
       mock.expects('_request').once().withArgs(
+        'GET', 'balance', null, callback);
+      bitx.getBalance(callback);
+      mock.verify();
+    });
+
+    it('should accept an asset argument and call _request with the correct parameters', function() {
+      mock.expects('_request').once().withArgs(
         'GET', 'balance', {asset: 'ZAR'}, callback);
       bitx.getBalance('ZAR', callback);
       mock.verify();